my_compliance_agent/pyproject.toml
[project]
name = "compliance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.38.23",
    "mcp-agent",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.3",
]

[tool.uv.sources]
mcp-agent = { workspace = true }
